Revert "Feature invite access"
