+++ apps/web/components/ui/Navbar.tsx
@@ -11,39 +11,17 @@ import {
   useColorMode,
 } from "@chakra-ui/react";
 import useUser from "@hooks/useUser";
-import { Role } from "@prisma/client";
 import { onLogout } from "@util/helpers";
 import axios from "axios";
 import { useRouter } from "next/router";
-import { RoleContext } from "pages/drives/[id]";
-import React, { useContext, useState } from "react";
-import {
-  ArrowNarrowLeft,
-  ChevronDown,
-  Coin,
-  File,
-  Logout,
-  Moon,
-  Sun,
-  User,
-} from "tabler-icons-react";
-import Invite from "./Invite";
-import InviteNotification from "./InviteNotification";
-import UserManagement from "./UserManagement";
+import React from "react";
+import { ArrowNarrowLeft, ChevronDown, Coin, File, Logout, Moon, Sun } from "tabler-icons-react";
 
 export default function Navbar() {
   const { colorMode, toggleColorMode } = useColorMode();
   const { mutateUser } = useUser();
   const router = useRouter();
-  const role = useContext(RoleContext);
-  const [isUserManagementModalOpen, setIsUserManagementModalOpen] = useState(false);
-  const handleOpenUserManagementModal = () => {
-    setIsUserManagementModalOpen(true);
-  };
 
-  const handleCloseUserManagementModal = () => {
-    setIsUserManagementModalOpen(false);
-  };
   return (
     <>
       <Flex
@@ -63,10 +41,7 @@ export default function Navbar() {
             Dashboard
           </Button>
         ) : null}
-
         <Box>
-          {router.route !== "/" && (role == Role.CREATOR || role == Role.ADMIN) && <Invite />}
-          <InviteNotification />
           <IconButton
             aria-label="toggle color theme"
             size="md"
@@ -80,20 +55,13 @@ export default function Navbar() {
               Actions
             </MenuButton>
             <MenuList>
-              {router.route !== "/" && (role == Role.CREATOR || role == Role.ADMIN) && (
-                <MenuItem height="8" icon={<User />} onClick={handleOpenUserManagementModal}>
-                  User Management
-                </MenuItem>
-              )}
               <MenuItem
-                height="8"
                 icon={<File />}
                 onClick={() => window.open("https://firefiles.app/docs", "_blank")}
               >
                 View Documentation
               </MenuItem>
               <MenuItem
-                height="8"
                 icon={<Coin />}
                 onClick={() => {
                   const url = "https://github.com/faisalsayed10/firefiles#sponsor-this-project";
@@ -103,7 +71,6 @@ export default function Navbar() {
                 Donate Us
               </MenuItem>
               <MenuItem
-                height="8"
                 icon={<Logout />}
                 onClick={async () => {
                   await onLogout();
@@ -119,7 +86,6 @@ export default function Navbar() {
         </Box>
       </Flex>
       <Divider />
-      <UserManagement isOpen={isUserManagementModalOpen} onClose={handleCloseUserManagementModal} />
     </>
   );
 }
