+++ apps/web/prisma/schema.prisma
@@ -9,43 +9,25 @@ datasource db {
 }
 
 model User {
-  id           String           @id @unique @default(uuid())
-  lastSignedIn DateTime         @map("last_signed_in")
-  createdAt    DateTime         @default(now()) @map("created_at")
-  email        String           @unique
-  verified     Boolean          @default(false)
-  plan         PricingPlan      @default(FREE)
-  drives       BucketsOnUsers[]
+  id           String      @id @unique @default(uuid())
+  lastSignedIn DateTime    @map("last_signed_in")
+  createdAt    DateTime    @default(now()) @map("created_at")
+  email        String      @unique
+  verified     Boolean     @default(false)
+  plan         PricingPlan @default(FREE)
+  drives       Drive[]
 }
 
 model Drive {
-  id        String           @id @unique @default(uuid())
-  createdAt DateTime         @default(now()) @map("created_at")
-  name      String
-  keys      String           @db.LongText
-  type      String
-  users    BucketsOnUsers[]
-}
-
-model BucketsOnUsers {
-  user      User     @relation(fields: [userId], references: [id])
-  userId    String
-  bucket    Drive    @relation(fields: [bucketId], references: [id])
-  bucketId  String
+  id        String   @id @unique @default(uuid())
   createdAt DateTime @default(now()) @map("created_at")
-  isPending Boolean  @default(true) @map("is_pending")
-  role      Role     @default(VIEWER)
+  name      String
+  keys      String   @db.LongText
+  type      String
+  User      User?    @relation(fields: [userId], references: [id])
+  userId    String?
 
-  @@id([userId, bucketId])
   @@index([userId])
-  @@index([bucketId])
-}
-
-enum Role {
-  CREATOR
-  ADMIN
-  VIEWER
-  EDITOR
 }
 
 enum PricingPlan {
