--- apps/web/util/emailSender/emailSender.ts
@@ -1,15 +1,0 @@
-export interface EmailSender {
-  sendLoginEmail(options: LoginEmailOptions): Promise<void>;
-  sendInvitationEmail(options: InvitationEmailOptions): Promise<void>;
-}
-
-export interface LoginEmailOptions {
-  email: string;
-  url: string;
-  token: string;
-}
-
-export interface InvitationEmailOptions {
-  email: string;
-  url: string;
-}
